+++ src/main/java/com/invadermonky/magicultureintegrations/integrations/astralsorcery/InitAstralSorcery.java
@@ -4,8 +4,10 @@ import com.invadermonky.magicultureintegrations.api.mods.IIntegrationModule;
 import com.invadermonky.magicultureintegrations.config.MIConfigTweaks;
 import com.invadermonky.magicultureintegrations.integrations.astralsorcery.block.BlockCrystalSorter;
 import com.invadermonky.magicultureintegrations.integrations.astralsorcery.events.ASEventSubscriber;
+import com.invadermonky.magicultureintegrations.integrations.astralsorcery.mods.ASSpartanWeaponry;
 import com.invadermonky.magicultureintegrations.registry.RegistrarMI;
 import com.invadermonky.magicultureintegrations.util.IntegrationList;
+import com.invadermonky.magicultureintegrations.util.ModIds;
 import net.minecraftforge.common.MinecraftForge;
 import org.jetbrains.annotations.NotNull;
 
@@ -15,7 +17,9 @@ public class InitAstralSorcery implements IIntegrationModule {
     public static BlockCrystalSorter crystal_sorter = new BlockCrystalSorter();
 
     @Override
-    public void buildModIntegrations() {}
+    public void buildModIntegrations() {
+        this.integrations.addIntegration(ModIds.spartan_weaponry, ASSpartanWeaponry.class);
+    }
 
     @Override
     public @NotNull IntegrationList getModIntegrations() {